requirements.txt
requests
qiniu
oss2
#aliyun-python-sdk-core  # used for python2.7
aliyun-python-sdk-core-v3  # used for python3
aliyun-python-sdk-cdn
Pillow
tornado
geoip2